+++ src/vue-router/index.js
@@ -353,12 +353,17 @@ function authPageGuard (to, from, next) {
 function redirectRouteGuard (to, from, next) {
   const isLoggedIn = store.getters[vuexTypes.isLoggedIn]
   const isAccountUnverified = store.getters[vuexTypes.isAccountUnverified]
+  const isKycRecoveryInProgress = store
+    .getters[vuexTypes.isKycRecoveryInProgress]
+
   if (isLoggedIn && !isAccountUnverified) {
     if (to.name === vueRoutes.app.name) {
       const isAccountCorporate = store.getters[vuexTypes.isAccountCorporate]
       const isBusinessToBrowse = store.getters[vuexTypes.isBusinessToBrowse]
       const isCustomerUiShown = store.getters[vuexTypes.isCustomerUiShown]
-      if (isAccountCorporate && !isCustomerUiShown) {
+      if (isKycRecoveryInProgress) {
+        next(vueRoutes.kycRecoveryManagement)
+      } else if (isAccountCorporate && !isCustomerUiShown) {
         next(vueRoutes.customers)
       } else if (isBusinessToBrowse) {
         next(vueRoutes.assetsExplore)